+++ CHANGES.rst
@@ -50,6 +50,9 @@ New Features
 
   - Added ``block_reduce`` and ``block_replicate`` functions. [#3453]
 
+  - extract array now offers different options to deal with array boundaries
+    [#3727]
+
 - ``astropy.stats``
 
 - ``astropy.table``